requirements.txt
fastapi >= '0.110.0'
uvicorn >= '0.30.0'
SQLAlchemy >= '2.0.0'
psycopg2 >= '2.9.0'

python-dotenv >= '1.0.0'